+++ xarray/tests/test_dataset.py
@@ -7815,7 +7815,10 @@ class TestDataset:
 # pytest tests — new tests should go here, rather than in the class.
 
 
-@pytest.mark.parametrize("test_elements", ([1, 2], np.array([1, 2]), DataArray([1, 2])))
+@pytest.mark.parametrize(
+    "test_elements",
+    ([1, 2], np.array([1, 2]), DataArray([1, 2]), {1, 2}, frozenset({1, 2})),
+)
 def test_isin(test_elements, backend) -> None:
     expected = Dataset(
         data_vars={